Fix remaining critical 64-bit portability issues
1. Plugin refCon pointer storage (CPluginDLL):
   - Changed mRefCon from long to intptr_t to safely store pointers
   - Added explicit casts at plugin API boundary
   - Prevents truncation on platforms where long != pointer size

2. ACAP/SASL protocol buffer lengths (CINETClient, CACAPClient):
   - Changed size_t to uint32_t for network protocol length fields
   - Ensures wire protocol compatibility between 32/64-bit systems
   - Fixed 4 instances in CACAPClient where size_t was written to network

3. AppleSingle/AppleDouble file format (CAFFilter):
   - Changed all unsigned long to uint32_t in file format structs
   - Fixed pointer arithmetic to use sizeof(uint32_t) consistently
   - Prevents file format corruption on 64-bit builds

4. BinHex file format (CBinHexFilter):
   - Changed sizeof(long) to sizeof(uint32_t) in pointer arithmetic
   - Ensures correct BinHex file decoding on 64-bit systems

5. Offline cache file seeks (CMailRecord):
   - Changed sizeof(unsigned long) to sizeof(uint32_t) for file seeks
   - Consistent with earlier fixes to mNextID (already uint32_t)
   - Prevents cache file corruption

All fixes maintain backward compatibility with 32-bit file formats while
ensuring correct operation on 64-bit LP64 systems (Linux x86_64, macOS).

+++ Plug-ins/Common/CPluginDLL.h
@@ -37,6 +37,8 @@
 #ifndef __PLUGIN_MULBERRY__
 #define __PLUGIN_MULBERRY__
 
+#include <stdint.h>
+
 // Typedefs
 
 #ifdef __cplusplus
@@ -115,7 +117,7 @@ public:
 	// Get information
 	virtual const SPluginInfo& GetInfo(void)					// Returns the info for the plug-in
 		{ return mInfo; }
-	virtual long GetRefCon(void)								// Returns refCon of plug-in
+	virtual intptr_t GetRefCon(void)							// Returns refCon of plug-in
 		{ return mRefCon; }
 
 	// Entry codes
@@ -134,7 +136,7 @@ public:
 protected:
 	LoggingCallbackProcPtr mLogging;						// Callback into application for logging
 	SPluginInfo mInfo;										// Information about plug-in
-	long mRefCon;											// Reference constant used by plugin: not used in app
+	intptr_t mRefCon;										// Reference constant used by plugin: not used in app
 
 	virtual void ErrorReport(long err_no, const char* errtxt, const char* func, const char* file, int lineno);
 	virtual void LogEntry(const char* text) const;			// Log some text
